Cleaned up code and removed unused state
Modernized several helpers and widgets with simpler syntax, removed
unused fields, dead code, and unnecessary renderer/body dependencies.
Simplified WSA reader/player state, removed unused BodyOrientation usage
from D2LeavesTracks, trimmed unused WorldRenderer storage from
D2SpriteWidget, and counted successfully imported original maps when they
were saved.

+++ OpenRA.Mods.D2/ImportData/D2MapSeed.cs
@@ -24,7 +24,7 @@ namespace OpenRA.Mods.D2.ImportData
 		{
 			get
 			{
-				return (uint)(seed[0])
+				return seed[0]
 					+ (uint)(seed[1] << 8)
 					+ (uint)(seed[2] << 16)
 					+ (uint)(seed[3] << 24);
